added asynchronous programming

+++ app/src/main/java/com/example/diabetesapp/code/GlucoseMeasurement.kt
@@ -3,6 +3,6 @@ package com.example.diabetesapp.code
 data class GlucoseMeasurement(
     var id: String = "",
     val value: Double = 0.0,
-    val time: Double = 0.0,
+    val time: Long = 0L,
     val userId: String = "" //żeby te dane były powiązane z userem
 )
